LUMEN Adjustments
