+++ src/Refne/src/main.jl
@@ -67,9 +67,9 @@ See also
 [`listrules`](@ref),
 [`rulemetrics`](@ref).
 """
-function refne(f, Xmin, Xmax; L=100, perc=1.0, max_depth=-1, n_subfeatures=-1, partial_sampling=0.7, min_samples_leaf=5, min_samples_split=2, min_purity_increase=0.0, seed=3, ott_mode = false)
+function refne(f, Xmin, Xmax; L=100, perc=1.0, max_depth=-1, n_subfeatures=-1, partial_sampling=0.7, min_samples_leaf=5, min_samples_split=2, min_purity_increase=0.0, seed=3, opt_mode = false)
 
-    if ott_mode == true
+    if opt_mode == true
         ddf = generate_univers_of_combinations_ott(Xmin, Xmax, L)
     else
         ddf = generate_univers_of_combinations(Xmin, Xmax, L)
