+++ src/lumen/utils/coreOptMode.jl
